fix: all issues + refactoring
