+++ packages/client/src/screens/scoreboard/globalScoreboard.screen.tsx
@@ -38,10 +38,89 @@ const MyGlobalScoreboardScreen = () => {
     refetch();
   }, [sortBy]);
 
+  const content = (
+    <div className="overflow-x-auto">
+      <table className="table table-zebra w-full">
+        {/* head */}
+        <thead>
+          <tr>
+            <th></th>
+            <th>Name</th>
+            <th>Score</th>
+            <th>Deviation (avg.)</th>
+            <th>Difficulty</th>
+          </tr>
+        </thead>
+        <tbody>
+          {/* row 1 */}
+          {data
+            ?.filter((game) => {
+              if (!sortBy) return true;
+              return game.difficulty.id === sortBy;
+            })
+            .sort((a, b) => b.score - a.score)
+            .map((game, index) => {
+              return (
+                <tr
+                  key={'game-' + (index + 1)}
+                  className={
+                    game.userId === settingsManager.getUserId()
+                      ? 'text-secondary'
+                      : ''
+                  }
+                >
+                  <th>
+                    {userHasWonGame(game) ? (
+                      <YouTooltip>{index + 1} </YouTooltip>
+                    ) : (
+                      index + 1
+                    )}
+                  </th>
+                  <td>
+                    {userHasWonGame(game) ? (
+                      <YouTooltip>
+                        {game?.name?.toUpperCase() || '???'}
+                      </YouTooltip>
+                    ) : (
+                      game?.name?.toUpperCase() || '???'
+                    )}
+                  </td>
+                  <td>
+                    {userHasWonGame(game) ? (
+                      <YouTooltip>{game.score} </YouTooltip>
+                    ) : (
+                      game.score
+                    )}
+                  </td>
+                  <td>
+                    {userHasWonGame(game) ? (
+                      <YouTooltip>
+                        {game.averageDeviation?.toFixed(2)}
+                        ms
+                      </YouTooltip>
+                    ) : (
+                      game.averageDeviation?.toFixed(2) + 'ms'
+                    )}
+                  </td>
+                  <td>
+                    {userHasWonGame(game) ? (
+                      <YouTooltip>{game.difficulty.name} </YouTooltip>
+                    ) : (
+                      game.difficulty.name
+                    )}
+                  </td>
+                </tr>
+              );
+            })}
+        </tbody>
+      </table>
+      {/* {isFetching && <div>Loading...</div>}
+  <div ref={targetRef}></div> */}
+    </div>
+  );
   if (isError) return <div>Error</div>;
   if (isLoading) return <div>Loading...</div>;
   if (data === null || data === undefined) return <div>Backend died.</div>;
-  if (!data.length) return <div>No Entries</div>;
 
   return (
     <div className="h-full flex flex-col px-2">
@@ -62,84 +141,7 @@ const MyGlobalScoreboardScreen = () => {
           </ul>
         </div>
       </div>
-      <div className="overflow-x-auto">
-        <table className="table table-zebra w-full">
-          {/* head */}
-          <thead>
-            <tr>
-              <th></th>
-              <th>Name</th>
-              <th>Score</th>
-              <th>Deviation (avg.)</th>
-              <th>Difficulty</th>
-            </tr>
-          </thead>
-          <tbody>
-            {/* row 1 */}
-            {data
-              ?.filter((game) => {
-                if (!sortBy) return true;
-                return game.difficulty.id === sortBy;
-              })
-              .sort((a, b) => b.score - a.score)
-              .map((game, index) => {
-                return (
-                  <tr
-                    key={'game-' + (index + 1)}
-                    className={
-                      game.userId === settingsManager.getUserId()
-                        ? 'text-secondary'
-                        : ''
-                    }
-                  >
-                    <th>
-                      {userHasWonGame(game) ? (
-                        <YouTooltip>{index + 1} </YouTooltip>
-                      ) : (
-                        index + 1
-                      )}
-                    </th>
-                    <td>
-                      {userHasWonGame(game) ? (
-                        <YouTooltip>
-                          {game?.name?.toUpperCase() || '???'}
-                        </YouTooltip>
-                      ) : (
-                        game?.name?.toUpperCase() || '???'
-                      )}
-                    </td>
-                    <td>
-                      {userHasWonGame(game) ? (
-                        <YouTooltip>{game.score} </YouTooltip>
-                      ) : (
-                        game.score
-                      )}
-                    </td>
-                    <td>
-                      {userHasWonGame(game) ? (
-                        <YouTooltip>
-                          {game.averageDeviation?.toFixed(2)}
-                          ms
-                        </YouTooltip>
-                      ) : (
-                        game.averageDeviation?.toFixed(2) + 'ms'
-                      )}
-                    </td>
-                    <td>
-                      {userHasWonGame(game) ? (
-                        <YouTooltip>{game.difficulty.name} </YouTooltip>
-                      ) : (
-                        game.difficulty.name
-                      )}
-                    </td>
-                  </tr>
-                );
-              })}
-          </tbody>
-        </table>
-        {/* {isFetching && <div>Loading...</div>}
-        <div ref={targetRef}></div> */}
-      </div>
+      {data.length ? content : 'No Entries'}
     </div>
   );
 };
