+++ packages/client/src/screens/scoreboard/personalScoreboard.screen.tsx
@@ -28,11 +28,51 @@ const MyPersonalScoreboardScreen = () => {
   const [sortBy, setSortBy] = useState<string | undefined>();
 
   //const [targetRef, isFetching] = useInfiniteScroll(fetchMoreData);
+  const content = (
+    <div className="overflow-x-auto">
+      <table className="table table-zebra w-full">
+        {/* head */}
+        <thead>
+          <tr>
+            <th></th>
+            <th>Name</th>
+            <th>Score</th>
+            <th>Deviation (avg.)</th>
+            <th>Difficulty</th>
+          </tr>
+        </thead>
+        <tbody>
+          {/* row 1 */}
+          {data
+            ?.filter((game) => {
+              if (!sortBy) return true;
+              return game.difficulty.id === sortBy;
+            })
+            .sort((a, b) => b.score - a.score)
+            .map((game, index) => {
+              return (
+                <tr key={'game-' + (index + 1)}>
+                  <th>{index + 1}</th>
+                  <td>{game?.name?.toUpperCase() || '???'}</td>
+                  <td>{game.score}</td>
+                  <td>
+                    {game.averageDeviation?.toFixed(2)}
+                    ms
+                  </td>
+                  <td>{game.difficulty?.name}</td>
+                </tr>
+              );
+            })}
+        </tbody>
+      </table>
+      {/* {isFetching && <div>Loading...</div>}
+  <div ref={targetRef}></div> */}
+    </div>
+  );
 
   if (isError) return <div>Error</div>;
   if (isLoading) return <div>Loading...</div>;
   if (data === null || data === undefined) return <div>Backend died.</div>;
-  if (!data.length) return <div>No Entries</div>;
 
   return (
     <div className="h-full flex flex-col px-2">
@@ -57,45 +97,12 @@ const MyPersonalScoreboardScreen = () => {
         </div>
       </div>
 
-      <div className="overflow-x-auto">
-        <table className="table table-zebra w-full">
-          {/* head */}
-          <thead>
-            <tr>
-              <th></th>
-              <th>Name</th>
-              <th>Score</th>
-              <th>Deviation (avg.)</th>
-              <th>Difficulty</th>
-            </tr>
-          </thead>
-          <tbody>
-            {/* row 1 */}
-            {data
-              ?.filter((game) => {
-                if (!sortBy) return true;
-                return game.difficulty.id === sortBy;
-              })
-              .sort((a, b) => b.score - a.score)
-              .map((game, index) => {
-                return (
-                  <tr key={'game-' + (index + 1)}>
-                    <th>{index + 1}</th>
-                    <td>{game?.name?.toUpperCase() || '???'}</td>
-                    <td>{game.score}</td>
-                    <td>
-                      {game.averageDeviation?.toFixed(2)}
-                      ms
-                    </td>
-                    <td>{game.difficulty?.name}</td>
-                  </tr>
-                );
-              })}
-          </tbody>
-        </table>
-        {/* {isFetching && <div>Loading...</div>}
-        <div ref={targetRef}></div> */}
-      </div>
+      {data?.filter((game) => {
+        if (!sortBy) return true;
+        return game.difficulty.id === sortBy;
+      }).length
+        ? content
+        : 'No Entries.'}
     </div>
   );
 };
