+++ packages/core/src/lib/settings-manager/default-settings-handler/decorators/rehydrate.middleware.ts
@@ -1,15 +1,26 @@
 import { Settings } from '../../../interfaces';
 import { Middleware } from '../../../middleware';
+import { DifficultyBuilder } from '../../modules';
 import { difficulties } from '../../modules/difficulty/difficulty';
 
 export const RehydrateSettingsMiddleware: Middleware<Settings> = (
   context,
   next
 ) => {
-  const hasDifficultyStrategy =
-    context.difficulty.key === 'DIFFICULTY_STRATEGY';
+  const hasDehydratedDifficulty = context.difficulty;
+  if (!hasDehydratedDifficulty) return next();
+
+  const hasDifficultyStrategy = context.difficulty.key === 'DIFFICULTY';
   if (hasDifficultyStrategy) {
-    context.difficulty = difficulties[context.difficulty.id];
+    const hasId = context.difficulty.id;
+    if (hasId) {
+      const difficultyFromId = difficulties[context.difficulty.id];
+      if (difficultyFromId) {
+        context.difficulty = difficulties[context.difficulty.id];
+      } else {
+        context.difficulty = new DifficultyBuilder().buildEasy();
+      }
+    }
   }
   return next();
 };
