+++ packages/core/src/lib/settings-manager/default-settings-handler/default-settings.handler.ts
@@ -1,14 +1,15 @@
 import { v4 as uuid } from 'uuid';
 import { Settings } from '../../interfaces';
 import { Middleware, MiddlewareHandler } from '../../middleware/middleware';
-import { DefaultColoring, easyDifficulty } from '../modules';
+import { DefaultColoring } from '../modules/coloring';
+import { DifficultyBuilder } from '../modules/difficulty/difficulty';
 
 export class DefaultSettingsHandlerImpl {
   private readonly settings: Settings;
 
   static readonly defaultSettings: Settings = {
     coloring: new DefaultColoring(),
-    difficulty: easyDifficulty,
+    difficulty: new DifficultyBuilder().buildEasy(),
     userId: uuid(),
     username: '',
   };
