+++ packages/core/src/lib/settings-manager/modules/difficulty/modules/index.ts
@@ -1,0 +1,4 @@
+export * from './fifty-fifty.difficulty';
+export * from './standard.difficulty';
+export * from './timer.difficulty';
+export * from './variable-deviation.difficulty';
